pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-recall"
version = "0.1.2"
description = "Git repo indexer + recall API (hybrid semantic + lexical search) with pgvector"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Repo Recall Contributors"}]
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "psycopg[binary]>=3.1.18",
  "pgvector>=0.2.5",
  "openai>=1.0.0",
  "gitpython>=3.1.43",
  "pathspec>=0.12.1",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "tenacity>=8.2.3",
  "httpx>=0.27.0",
  "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "ruff>=0.5.0",
  "types-python-dateutil",
]

[project.scripts]
repo-recall = "repo_recall.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["repo_recall"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]


[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
repo_recall = ["schema.sql", "ui/templates/*.html", "ui/static/*"]
